完成了cache系统，并使PySimpleGUI正式成为引擎的一部分

+++ linpg/basic/mixer.py
@@ -65,7 +65,6 @@ class SoundManagement(AbstractSoundManager):
 
 # 获取视频的音频 （返回路径）
 def _split_audio_from_video(input_path: str, audio_type: object = "ogg") -> str:
-    output_folder: str = os.path.dirname(input_path)
     # 产生不重名的output文件名称
     output_file_name_t: str = os.path.basename(input_path).replace(".", "_") + "{0}.{1}"
     output_file_name: str
@@ -77,7 +76,7 @@ def _split_audio_from_video(input_path: str, audio_type: object = "ogg") -> str:
         else:
             index += 1
     # 生成output路径
-    output_path: str = os.path.join(output_folder, output_file_name)
+    output_path: str = os.path.join(Cache.generate_folder_path(), output_file_name)
     try:
         # 生成视频文件
         VideoConverter.convert_from_video_to_audio(input_path, output_path)
@@ -106,10 +105,22 @@ class Sound:
 
     # 从一个视频中加载音效
     @staticmethod
-    def load_from_video(path: str, volume: float = 1.0) -> PG_Sound:
+    def load_from_video(path: str, volume: float = 1.0, cache_key: Optional[str] = None) -> PG_Sound:
+        # 如果给定了cache_key，则先尝试从缓存中读取音乐文件
+        if cache_key is not None and len(cache_key) > 0 and Cache.match(cache_key, path) is True:
+            try:
+                return Sound.load(Cache.get_cache_path(cache_key), volume)
+            except Exception:
+                pass
+        # 如果读取失败或者没有缓存key或者match失败，则应根据给定的路径生成音乐文件并返回
         path_of_sound: str = _split_audio_from_video(path)
-        sound_audio = Sound.load(path_of_sound, volume)
-        os.remove(path_of_sound)
+        sound_audio: PG_Sound = Sound.load(path_of_sound, volume)
+        # 如果给了缓存key，则应该生成缓存联系并保留缓存文件
+        if cache_key is not None and len(cache_key) > 0:
+            Cache.new(cache_key, path, path_of_sound)
+        # 如果没有缓存key，则删除缓存文件
+        else:
+            os.remove(path_of_sound)
         return sound_audio
 
     # 播放音效
